refactor(validator): user, company schema 파일명 변경
